requirements.txt
Django==2.2.10
djangorestframework==3.12.2
pytz==2021.1
PyYAML==5.4.1
sqlparse==0.4.1
uritemplate==3.0.1
